+++ pylabrobot/liquid_handling/liquid_handler_tests.py
@@ -35,6 +35,7 @@ from pylabrobot.resources.errors import (
   NoTipError,
 )
 from pylabrobot.resources.hamilton import HTF, STF, STARLetDeck
+from pylabrobot.resources.opentrons.reservoirs import agilent_1_reservoir_290ml
 from pylabrobot.resources.utils import create_ordered_items_2d
 from pylabrobot.resources.volume_tracker import (
   set_cross_contamination_tracking,
@@ -45,14 +46,14 @@ from pylabrobot.resources.well import Well
 from . import backends
 from .liquid_handler import LiquidHandler, OperationCallback
 from .standard import (
-  Aspiration,
-  AspirationPlate,
-  Dispense,
-  DispensePlate,
   Drop,
   DropTipRack,
   GripDirection,
+  MultiHeadAspirationPlate,
+  MultiHeadDispensePlate,
   Pickup,
+  SingleChannelAspiration,
+  SingleChannelDispense,
 )
 
 
@@ -61,8 +62,8 @@ def _make_asp(
   vol: float,
   tip: Any,
   offset: Coordinate = Coordinate.zero(),
-) -> Aspiration:
-  return Aspiration(
+) -> SingleChannelAspiration:
+  return SingleChannelAspiration(
     resource=r,
     volume=vol,
     tip=tip,
@@ -79,8 +80,8 @@ def _make_disp(
   vol: float,
   tip: Any,
   offset: Coordinate = Coordinate.zero(),
-) -> Dispense:
-  return Dispense(
+) -> SingleChannelDispense:
+  return SingleChannelDispense(
     resource=r,
     volume=vol,
     tip=tip,
@@ -781,7 +782,7 @@ class TestLiquidHandlerCommands(unittest.IsolatedAsyncioTestCase):
         "command": "aspirate96",
         "args": (),
         "kwargs": {
-          "aspiration": AspirationPlate(
+          "aspiration": MultiHeadAspirationPlate(
             wells=self.plate.get_all_items(),
             volume=10.0,
             tips=ts,
@@ -800,7 +801,7 @@ class TestLiquidHandlerCommands(unittest.IsolatedAsyncioTestCase):
         "command": "dispense96",
         "args": (),
         "kwargs": {
-          "dispense": DispensePlate(
+          "dispense": MultiHeadDispensePlate(
             wells=self.plate.get_all_items(),
             volume=10.0,
             tips=ts,
@@ -1008,6 +1009,11 @@ class TestLiquidHandlerCommands(unittest.IsolatedAsyncioTestCase):
     self.assertEqual(well_a2.tracker.liquids, [(None, 10)])
 
     set_volume_tracking(enabled=False)
+
+  async def test_aspirate_single_reservoir(self):
+    reagent_reservoir = agilent_1_reservoir_290ml(name="reservoir")
+    await self.lh.pick_up_tips96(self.tip_rack)
+    await self.lh.aspirate96(reagent_reservoir.get_item("A1"), volume=100)
 
 
 class TestLiquidHandlerVolumeTracking(unittest.IsolatedAsyncioTestCase):