+++ pylabrobot/liquid_handling/standard.py
@@ -49,7 +49,7 @@ class DropTipRack:
 
 
 @dataclass(frozen=True)
-class Aspiration:
+class SingleChannelAspiration:
   resource: Container
   offset: Coordinate
   tip: Tip
@@ -61,7 +61,7 @@ class Aspiration:
 
 
 @dataclass(frozen=True)
-class Dispense:
+class SingleChannelDispense:
   resource: Container
   offset: Coordinate
   tip: Tip
@@ -73,7 +73,7 @@ class Dispense:
 
 
 @dataclass(frozen=True)
-class AspirationPlate:
+class MultiHeadAspirationPlate:
   wells: List[Well]
   offset: Coordinate
   tips: List[Tip]
@@ -85,7 +85,7 @@ class AspirationPlate:
 
 
 @dataclass(frozen=True)
-class DispensePlate:
+class MultiHeadDispensePlate:
   wells: List[Well]
   offset: Coordinate
   tips: List[Tip]
@@ -97,7 +97,7 @@ class DispensePlate:
 
 
 @dataclass(frozen=True)
-class AspirationContainer:
+class MultiHeadAspirationContainer:
   container: Container
   offset: Coordinate
   tips: List[Tip]
@@ -109,7 +109,7 @@ class AspirationContainer:
 
 
 @dataclass(frozen=True)
-class DispenseContainer:
+class MultiHeadDispenseContainer:
   container: Container
   offset: Coordinate
   tips: List[Tip]
@@ -155,4 +155,4 @@ class ResourceDrop:
   rotation: float
 
 
-PipettingOp = Union[Pickup, Drop, Aspiration, Dispense]
+PipettingOp = Union[Pickup, Drop, SingleChannelAspiration, SingleChannelDispense]